requirements.txt
# Core dependencies
google-generativeai
langchain
langchain-google-genai
openai
langchain-openai

# Data validation and structuring
pydantic

# Ghost CMS API communication
requests
PyJWT

# HTML/Markdown processing
markdown
markdownify
beautifulsoup4

# Environment variable management
python-dotenv